Addressed code review points
