+++ pytorch_lightning/trainer/connectors/precision_connector.py
@@ -16,9 +16,10 @@ from typing import Optional
 from pytorch_lightning import _logger as log
 from pytorch_lightning.plugins.apex import ApexPlugin
 from pytorch_lightning.plugins.native_amp import NativeAMPPlugin
-from pytorch_lightning.plugins.sharded_native_amp_plugin import ShardedNativeAMPPlugin, FAIRSCALE_AMP_AVAILABLE
+from pytorch_lightning.plugins.sharded_native_amp_plugin import ShardedNativeAMPPlugin
 from pytorch_lightning.plugins.sharded_plugin import DDPShardedPlugin
-from pytorch_lightning.utilities import APEX_AVAILABLE, NATIVE_AMP_AVALAIBLE, AMPType, rank_zero_warn
+from pytorch_lightning.utilities import APEX_AVAILABLE, NATIVE_AMP_AVALAIBLE, AMPType, rank_zero_warn, \
+    FAIRSCALE_AVAILABLE
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
@@ -60,7 +61,7 @@ class PrecisionConnector:
             else:
                 self.trainer.amp_backend = AMPType.NATIVE
                 if plugins and self._sharded_in_plugins(plugins):
-                    if not FAIRSCALE_AMP_AVAILABLE:
+                    if not FAIRSCALE_AVAILABLE:
                         raise MisconfigurationException('Sharded DDP Plugin requires Fairscale to be installed.')
                     log.info('Using Sharded 16bit plugin.')
                     self.backend = ShardedNativeAMPPlugin(self.trainer)
